requirements.txt

playwright==1.39.0
pytest==7.4.0
python-dotenv==1.0.0
pytest-html==3.2.0
